+++ shared/include/shared/map.h
@@ -2,13 +2,12 @@
 #define MAP_H
 
 #include <stdbool.h>
-#include <stdint.h>
 
 #define MAX_MOBS 20
 
 struct tile
 {
-    int64_t gid;
+    size_t gid;
     bool h_flip;
     bool v_flip;
     bool d_flip;
@@ -16,7 +15,7 @@ struct tile
 
 struct mob
 {
-    int64_t gid;
+    size_t gid;
 
     float origin_x;
     float origin_y;
@@ -53,6 +52,6 @@ void map_unload(struct map *map);
 void map_update(struct map *map, float delta_time);
 
 struct tile *map_get_tile(struct map *map, size_t x, size_t y);
-struct tileset *map_get_tileset(struct map *map, int64_t gid);
+struct tileset *map_get_tileset(struct map *map, size_t gid);
 
 #endif
