+++ shared/include/shared/tileset.h
@@ -14,7 +14,7 @@ struct tileset
     size_t index;
 
     char *filename;
-    int64_t first_gid;
+    size_t first_gid;
 
     size_t columns;
     char *image;
@@ -24,6 +24,6 @@ struct tileset
 void tileset_load(struct tileset *tileset, char *filename);
 void tileset_unload(struct tileset *tileset);
 
-struct tile_data *tileset_get_tile_data(struct tileset *tileset, int64_t gid);
+struct tile_data *tileset_get_tile_data(struct tileset *tileset, size_t gid);
 
 #endif
